Ajoute des options de typechecking pour TS (#1156)
